requirements.txt
streamlit==1.43.0
groq==0.11.0
mcp==1.6.0

pandas==2.2.0
numpy==1.26.4

python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0

google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

notion-client==2.2.1

tavily-python==0.3.3

python-dateutil==2.9.0

watchdog==4.0.0
aiohttp==3.9.3